Started some code cleanup
